feat: restructure flags to retrieve closed PRs by default
When I initially wrote this, for some reason I didn't need to export
the closed PRs, so I just hardcoded it that way.

Change the default behaviour to always pull in closed PRs (which is
consistent with the default of pulling in closed issues). Also add new
flags --no-closed-prs, --no-closed-issues, --no-prs and --no-issues,
to selectively disable parts of the export.