fix: allow GITHUB_API_URL to point to api.github.com
A change yesterday added support for setting GITHUB_API_URL to override
the base graphql API URL. This expects to include the path to point to
the graphql element. But, the nightly Github action failed yesterday
because GITHUB_API_URL is set automatically in Github actions to
https://api.github.com.

We now handle this case automatically, for convenience: if the url
points to api.github.com then we'll resolve it to the graphql url,
but if it's something custom then we'll expect it to point to the
graphql location explicitly.